requirements.txt
Django>=5.0
rdflib>=7.0
djangorestframework>=3.14
django-filter>=23.0
pandas>=2.0
pyarrow>=14.0
markdown>=3.5
psycopg2-binary>=2.9.9
requests>=2.31